+++ CHANGELOG.md
@@ -15,7 +15,9 @@ UNRELEASED
 * Changed: oda v12.1.1
 * Added: polarisations to mapping
 * Fixed: use of REACT_APP_USE_LOCAL_DATA for local development
+* Changed: to show notification when there is an senscalc error after mapping target-observation
 * Changed: Added validation to target fields 
+
 
 8.0.0
 ----------